Updated security patches and removed redundancies
Updated security patches and removed redundancies

+++ backend/app/models/client_project.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Float, Text
+from sqlalchemy import Column, Integer, String, DateTime, Float, Text, Numeric
 from datetime import datetime
 from app.core.database import Base
 
@@ -13,8 +13,8 @@ class ClientProject(Base):
     execution_status = Column(String, default="pending")
     delivery_status = Column(String, default="pending")
     estimated_delivery = Column(DateTime, nullable=True)
-    cost = Column(Float, default=0.0)
-    amount_paid = Column(Float, default=0.0)
-    amount_pending = Column(Float, default=0.0)
+    cost = Column(Numeric(12, 2), default=0)
+    amount_paid = Column(Numeric(12, 2), default=0)
+    amount_pending = Column(Numeric(12, 2), default=0)
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
