+++ backend/app/models/estimate.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Text, Boolean
+from sqlalchemy import Column, Integer, String, Float, DateTime, Text, Boolean, Numeric
 from datetime import datetime
 from app.core.database import Base
 
@@ -9,9 +9,9 @@ class Estimate(Base):
     client_id = Column(Integer, index=True)
     client_name = Column(String)
     description = Column(Text)
-    material_cost = Column(Float, default=0.0)
-    labor_cost = Column(Float, default=0.0)
-    total_cost = Column(Float, default=0.0)
+    material_cost = Column(Numeric(12, 2), default=0)
+    labor_cost = Column(Numeric(12, 2), default=0)
+    total_cost = Column(Numeric(12, 2), default=0)
     status = Column(String, default="draft")
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
