+++ backend/app/models/salary_slip.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Float
+from sqlalchemy import Column, Integer, String, DateTime, Float, Numeric
 from datetime import datetime
 from app.core.database import Base
 
@@ -9,10 +9,10 @@ class SalarySlip(Base):
     employee_id = Column(Integer, index=True)
     month = Column(String)
     year = Column(String)
-    basic_salary = Column(Float, default=0.0)
-    allowances = Column(Float, default=0.0)
-    deductions = Column(Float, default=0.0)
-    net_salary = Column(Float, default=0.0)
+    basic_salary = Column(Numeric(12, 2), default=0)
+    allowances = Column(Numeric(12, 2), default=0)
+    deductions = Column(Numeric(12, 2), default=0)
+    net_salary = Column(Numeric(12, 2), default=0)
     status = Column(String, default="pending")
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
