+++ backend/app/schemas/client_project.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel
 from typing import Optional
 from datetime import datetime
@@ -6,7 +7,7 @@ class ClientProjectBase(BaseModel):
     client_id: int
     project_name: str
     description: Optional[str] = None
-    cost: float
+    cost: Decimal
 
 class ClientProjectCreate(ClientProjectBase):
     pass
@@ -18,17 +19,17 @@ class ClientProjectUpdate(BaseModel):
     execution_status: Optional[str] = None
     delivery_status: Optional[str] = None
     estimated_delivery: Optional[datetime] = None
-    cost: Optional[float] = None
-    amount_paid: Optional[float] = None
-    amount_pending: Optional[float] = None
+    cost: Optional[Decimal] = None
+    amount_paid: Optional[Decimal] = None
+    amount_pending: Optional[Decimal] = None
 
 class ClientProjectResponse(ClientProjectBase):
     id: int
     design_status: str
     execution_status: str
     delivery_status: str
-    amount_paid: float
-    amount_pending: float
+    amount_paid: Decimal
+    amount_pending: Decimal
     estimated_delivery: Optional[datetime]
     created_at: datetime
     updated_at: datetime
