+++ backend/app/schemas/estimate.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel
 from typing import Optional
 from datetime import datetime
@@ -6,8 +7,8 @@ class EstimateBase(BaseModel):
     client_id: int
     client_name: str
     description: str
-    material_cost: float
-    labor_cost: float
+    material_cost: Decimal
+    labor_cost: Decimal
 
 class EstimateCreate(EstimateBase):
     pass
@@ -15,13 +16,13 @@ class EstimateCreate(EstimateBase):
 class EstimateUpdate(BaseModel):
     client_name: Optional[str] = None
     description: Optional[str] = None
-    material_cost: Optional[float] = None
-    labor_cost: Optional[float] = None
+    material_cost: Optional[Decimal] = None
+    labor_cost: Optional[Decimal] = None
     status: Optional[str] = None
 
 class EstimateResponse(EstimateBase):
     id: int
-    total_cost: float
+    total_cost: Decimal
     status: str
     created_at: datetime
     updated_at: datetime
