+++ backend/app/schemas/salary_slip.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel
 from typing import Optional
 from datetime import datetime
@@ -6,22 +7,22 @@ class SalarySlipBase(BaseModel):
     employee_id: int
     month: str
     year: str
-    basic_salary: float
-    allowances: float = 0.0
-    deductions: float = 0.0
+    basic_salary: Decimal
+    allowances: Decimal = 0
+    deductions: Decimal = 0
 
 class SalarySlipCreate(SalarySlipBase):
     pass
 
 class SalarySlipUpdate(BaseModel):
-    basic_salary: Optional[float] = None
-    allowances: Optional[float] = None
-    deductions: Optional[float] = None
+    basic_salary: Optional[Decimal] = None
+    allowances: Optional[Decimal] = None
+    deductions: Optional[Decimal] = None
     status: Optional[str] = None
 
 class SalarySlipResponse(SalarySlipBase):
     id: int
-    net_salary: float
+    net_salary: Decimal
     status: str
     created_at: datetime
     updated_at: datetime
